Abtract
Automated segmentation of brain tumors in 3D MRI scans is crucial for efficient diagnosis, follow-up, and therapy planning. Manual delineation methods are resource-intensive and prone to human errors. In addressing this challenge, we propose a classification method to categorize brain tumors into three distinct classes. This approach aims to streamline the diagnostic process, reducing the reliance on time-consuming and labor-intensive manual techniques. By leveraging automated classification, clinicians can enhance their understanding of tumor symptoms and locations, leading to more efficient and accurate diagnosis and treatment planning. This not only accelerates the overall process but also contributes to a more effective and accessible healthcare system, particularly in the realm of neurological disorders.